StoreManager interface shouldn't depend upon underlying storage. Fix the disk available api and move ServerErrorCode to ambry-api.
