+++ ambry-api/src/main/java/com.github.ambry/server/ServerErrorCode.java
@@ -11,7 +11,7 @@
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  */
-package com.github.ambry.commons;
+package com.github.ambry.server;
 
 /**
  * The error codes that the server returns on a failed request
