+++ ambry-commons/src/test/java/com.github.ambry.commons/ResponseHandlerTest.java
@@ -24,6 +24,7 @@ import com.github.ambry.network.ConnectionPoolTimeoutException;
 import com.github.ambry.network.NetworkClientErrorCode;
 import com.github.ambry.router.RouterErrorCode;
 import com.github.ambry.router.RouterException;
+import com.github.ambry.server.ServerErrorCode;
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.SocketException;
