+++ ambry-server/src/integration-test/java/com.github.ambry.server/ServerHardDeleteTest.java
@@ -19,7 +19,6 @@ import com.github.ambry.clustermap.MockClusterMap;
 import com.github.ambry.clustermap.PartitionId;
 import com.github.ambry.commons.BlobId;
 import com.github.ambry.commons.CommonTestUtils;
-import com.github.ambry.commons.ServerErrorCode;
 import com.github.ambry.config.VerifiableProperties;
 import com.github.ambry.messageformat.BlobData;
 import com.github.ambry.messageformat.BlobProperties;
