+++ ambry-server/src/test/java/com.github.ambry.server/AmbryRequestsTest.java
@@ -27,7 +27,6 @@ import com.github.ambry.clustermap.ReplicaType;
 import com.github.ambry.commons.BlobId;
 import com.github.ambry.commons.CommonTestUtils;
 import com.github.ambry.commons.ErrorMapping;
-import com.github.ambry.commons.ServerErrorCode;
 import com.github.ambry.commons.ServerMetrics;
 import com.github.ambry.config.ClusterMapConfig;
 import com.github.ambry.config.DiskManagerConfig;
